assistant/requirements.txt
python-dotenv==1.0.0
langchain==0.1.0
llama_cpp_python
sentence-transformers==2.3.1
gradio==4.18.0
chromadb==0.4.22